"Implement login message display state in App component"
In response to the need for a login message display, a new state variable 'isLoginMessageDisplayed' has been added to the App component. This variable is initially set to false and is updated to true when the login is successful. 

A condition has been added to check if 'isLoginMessageDisplayed' is false and the user is logged in. If true, it sets 'isLoginMessageDisplayed' to true and displays a login message. If 'isLoginMessageDisplayed' is true, it does not display the login message. 

A new method binding for 'handleLogout' has been added in the constructor. A new method 'handleLogout' has been added that sets 'isLoginMessageDisplayed' back to false when a user logs out. 

This implementation differs from the previous one as it now provides a mechanism to control the display of the login message based on the login state and the 'isLoginMessageDisplayed' state.