"Update App component and Spotify utility for user info handling"
In this commit, several changes have been made to the App component and Spotify utility to improve the handling of user information. 

The initial state values of 'spotifyAvatar' and 'spotifyUsername' in the App component have been changed from empty strings to null. This change was motivated by the need to accurately represent the absence of these values in the state.

A new method 'checkUserInfo' has been added to the App component. This method checks if the username and avatar exist in the state and, if they don't, calls the 'handleLogin' method to fetch these details from the Spotify API. This method is called in the 'componentDidMount' lifecycle method.

The render method of the App component has been updated to conditionally render the username and avatar based on their existence in the state. If the 'spotifyUsername' is not available, 'Guest' will be displayed. If the 'spotifyAvatar' is not available, the img tag for the avatar will not be rendered.

In the Spotify utility, a check for 'response.ok' has been added in the fetch request. If the response is not ok, the function will return null. This change was made to handle failed fetch requests more gracefully.