"Add Spotify user info to Sidebar"
In response to the need for displaying the user's Spotify username and avatar in the Sidebar, several changes were made across the codebase. 

In src/util/Spotify.js, a new method getUserInfo() was added to the Spotify object. This method fetches the user's information from the Spotify API and returns an object containing the username and avatar URL. If the user's information cannot be retrieved, it returns an object with null values for the username and avatar.

In src/components/App/App.js, 'spotifyUsername' and 'spotifyAvatar' fields were added to the state in the constructor. The 'handleLogin' method was modified to call 'Spotify.getUserInfo' and store the retrieved username and avatar in the state. The render method was updated to display the Spotify username and avatar from the state if the user is logged in. If the user is not logged in, it will display the default values. A new img tag was added to display the Spotify avatar and a new h1 tag to display the Spotify username. These are placed above the 'Provide Feedback' button in the Sidebar div. The src for the img tag is set to this.state.spotifyAvatar or a default image if the avatar is null. The content for the h1 tag is set to this.state.spotifyUsername or a default string if the username is null.

In src/components/App/App.css, CSS rules were added for .username and .avatar inside .Sidebar. The .username has a font size of 1.2rem and a margin of 1rem. The .avatar has a width of 70%, a border-radius of 50% to make it circular, and centered with margin auto. Both .username and .avatar have pointer-events set to none to make them non-clickable. A media query was added to hide the Spotify username and avatar in the Sidebar for non-desktop views.

These changes enhance the user experience by providing a more personalized interface when the user is logged in.