"Update color of SearchBar input field to green"
This commit changes the color property of the .SearchBar input selector in the SearchBar.css file. The color has been updated from #010c3f to green. This change also applies to the media query for max-width: 600px.

The motivation for this change was to improve the visibility and aesthetic of the text in the search bar. The previous color (#010c3f) was deemed too dark and was replaced with a more vibrant green color. This change only affects the color property of the .SearchBar input selector and does not impact any other CSS rules or properties.