"Refactor fetch calls to use SpotifyApi.get in Spotify.js"
This commit replaces all instances of fetch with SpotifyApi.get in the Spotify.js utility file. The motivation for this change was to refactor the existing functions to use the Spotify SDK instead of the fetch function for making API requests. This new implementation is more efficient and aligns with the best practices for interacting with the Spotify API. 

The changes in this commit will also require updates to the App, SearchResults, Playlist, Track, and SearchBar components to use the new functions from the Spotify.js utility file. These updates will be made in subsequent commits.