"Update Playlist.js to use Spotify SDK"
In this commit, the Playlist.js file has been updated to use the Spotify SDK. The Spotify SDK has been imported and the savePlaylist method has been added and bound in the constructor. The onClick event of the SAVE TO SPOTIFY button has been changed to use the savePlaylist method. 

The motivation for these changes was to replace the existing fetch-based API calls with the corresponding methods provided by the Spotify SDK. This new implementation differs from the previous one as it now uses the Spotify SDK to save the playlist, which is more efficient and reliable.