"Replace fetch calls with SpotifyWebApi methods in Spotify.js"
In response to tasks related to improving the efficiency and readability of our codebase, we have replaced all fetch calls in the Spotify.js file with corresponding methods from the SpotifyWebApi. 

The motivation for this change was to streamline our code and make use of the SpotifyWebApi's built-in methods, which are more efficient and easier to read than traditional fetch calls. 

The new implementation differs from the previous one in that it no longer uses fetch calls to interact with the Spotify API. Instead, it uses the following SpotifyWebApi methods: getMe, searchTracks, getRecommendations, createPlaylist, and addTracksToPlaylist. 

This change should make our code cleaner and easier to maintain, as well as potentially improving performance by using the more efficient SpotifyWebApi methods.