"Import updated Spotify.js in Playlist.js"
This commit includes the necessary changes to import the updated Spotify.js file in the Playlist.js component. The motivation for this change was to ensure that the Playlist component has access to the latest features and updates included in the Spotify.js file. This differs from the previous implementation where the older version of Spotify.js was being used. By importing the updated file, we ensure that our Playlist component is up-to-date and can utilize all the functionalities provided by Spotify.js.