"Import updated Spotify.js in SearchResults.js"
This commit includes the addition of the import statement for the updated Spotify.js file in the SearchResults.js file. The motivation for this change was to ensure that the SearchResults component has access to the latest functionalities provided by the Spotify.js file. This differs from the previous implementation where the outdated Spotify.js file was being used. By importing the updated file, we ensure that the SearchResults component is always in sync with the latest changes in the Spotify.js file.