"Import updated Spotify.js in TrackList.js"
This commit adds the import statement for the updated Spotify.js file in the TrackList.js component. The motivation for this change was to ensure that the TrackList component has access to the latest functionalities provided by the Spotify.js file. This differs from the previous implementation where the older version of Spotify.js was being used. This update ensures that the TrackList component is in sync with the latest changes in the Spotify.js file.