"Integrate Spotify SDK in src/util/Spotify.js"
This commit includes the addition of the import statement for the Spotify SDK at the top of the src/util/Spotify.js file. Additionally, the Spotify object has been initialized using the SpotifyWebApi constructor.

The motivation for this change was to enable the use of Spotify's functionalities within our application. This was a necessary step in the tasks related to the integration of Spotify's services.

The previous implementation did not include the Spotify SDK, thus limiting the application's capabilities. With this change, we can now leverage Spotify's features to enhance our application's functionality. 

This commit represents a significant step towards the completion of several tasks related to Spotify integration.