"Refactor getUserInfo function in Spotify.js"
In the src/util/Spotify.js file, the getUserInfo function has been refactored. The previous fetch call has been replaced with the spotifyApi.getMe() method from the Spotify SDK. This change was motivated by the need to streamline the code and make use of the methods provided by the SDK, which are more efficient and reliable.

Additionally, the SpotifyWebApi object has been instantiated and the access token has been set for it. This is a necessary step to make the spotifyApi.getMe() method work correctly. This differs from the previous implementation where the access token was not explicitly set for the SpotifyWebApi object.

These changes were inspired by several tasks related to improving the backend services and enhancing the overall functionality of the application.