"Update getAccessToken function in Spotify.js"
Updated the getAccessToken function in src/util/Spotify.js to use the Spotify SDK's clientCredentialsGrant method for authentication. This change was motivated by the need to improve the efficiency and security of our authentication process. 

Previously, the getAccessToken function was using a less secure and less efficient method for authentication. The new implementation retrieves an access token from Spotify's API using the client ID and secret, which is more secure and efficient. The access token is then set for the Spotify API instance and stored in the accessToken variable. This change ensures that our application is using the most up-to-date and secure methods for interacting with Spotify's API.