"Add savePlaylist function to Spotify object in src/util/Spotify.js"
This commit introduces a new function, savePlaylist, to the Spotify object in the src/util/Spotify.js file. The function is designed to streamline the process of creating a new playlist and adding tracks to it on Spotify.

The motivation for this change was to simplify the process of creating and populating a new playlist on Spotify. Previously, these steps had to be performed separately, which was less efficient and could lead to errors.

The savePlaylist function first checks if a name and trackUris are provided. If not, it returns. It then gets the access token and sets the headers for the Spotify API request. It retrieves the user's Spotify ID using the getMe method from the Spotify API. Then, it creates a new playlist with the provided name using the createPlaylist method from the Spotify API. Finally, it adds the tracks to the newly created playlist using the addTracksToPlaylist method from the Spotify API.

This new implementation is more efficient and less error-prone than the previous approach. It also provides a more streamlined user experience by allowing the creation and population of a new playlist in a single step.